+++ dsr_bridge/CHANGELOG.rst
@@ -11,6 +11,7 @@ X.Y.Z (DD-MM-YYYY)
 * Converted to composables.
 * Move launch to bringup.
 * Move some functions to dsr_util::helpers.
+* Added function to check if the edge are lost in bridge.
 
 1.0.0 (04-03-2024)
 ------------------